% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd_l2svm.R
\name{create_l2svm_generator}
\alias{create_l2svm_generator}
\title{Create Data Generator for L2-SVM}
\usage{
create_l2svm_generator(
  mean_1 = c(1, 1, 1, 0, 0),
  mean_2 = c(0, 0, 1, 1, 1),
  mixture_weights = c(0.2, 0.8),
  Sigma = diag(rep(0.5, 5))
)
}
\arguments{
\item{mean_1}{Mean vector for first component}

\item{mean_2}{Mean vector for second component}

\item{mixture_weights}{Mixture weights}

\item{Sigma}{Covariance matrix}
}
\value{
Function that generates single data points when called
}
\description{
Creates a data generation function for L2-SVM based on specified mixture
distribution parameters. This is a convenience function for creating data
generators to use with SGD functions.
}
\examples{
# Create a data generator
data_gen <- create_l2svm_generator()

# Generate some data points
point1 <- data_gen()
point2 <- data_gen()

}
\keyword{internal} 