% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pra_logistic.R
\name{create_logistic_generator}
\alias{create_logistic_generator}
\title{Create Data Generator for Logistic Regression}
\usage{
create_logistic_generator(theta)
}
\arguments{
\item{theta}{True coefficient vector (parameter of the logistic model)}
}
\value{
Function that generates single data points when called
}
\description{
Creates a data generation function for logistic regression based on specified
true coefficients theta. This is a convenience function for creating data
generators to use with inference functions.
}
\examples{
# Create a data generator
theta <- c(1, -0.5, 0.3)
data_gen <- create_logistic_generator(theta)

# Generate some data points
point1 <- data_gen()
point2 <- data_gen()

}
\keyword{internal} 