% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd_l2svm.R
\name{generate_l2svm_data}
\alias{generate_l2svm_data}
\title{Generate L2-SVM Data}
\usage{
generate_l2svm_data(
  mean_1 = c(1, 1, 1, 0, 0),
  mean_2 = c(0, 0, 1, 1, 1),
  mixture_weights = c(0.2, 0.8),
  Sigma = diag(rep(0.5, 5))
)
}
\arguments{
\item{mean_1}{Mean vector for first component (default = c(1, 1, 1, 0, 0))}

\item{mean_2}{Mean vector for second component (default = c(0, 0, 1, 1, 1))}

\item{mixture_weights}{Mixture weights (default = c(0.2, 0.8))}

\item{Sigma}{Covariance matrix (default = diag(rep(0.5, 5)))}
}
\value{
List containing:
\item{X}{Design matrix (1 x d)}
\item{y}{Binary response (scalar) with value in {-1, +1}}
}
\description{
Generates (X, y) pairs for L2-SVM where features come from a mixture of two
multivariate normal distributions with specified means and covariance.
}
\examples{
# Generate L2-SVM data point
set.seed(123)
data <- generate_l2svm_data()
print(cbind(data$X, data$y))

}
\keyword{internal} 