% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pra_logistic.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\usage{
sigmoid(z)
}
\arguments{
\item{z}{Input vector or scalar}
}
\value{
Sigmoid transformation of z
}
\description{
Computes the sigmoid function.
}
\keyword{internal} 