% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_generators.R
\name{ske_iid}
\alias{ske_iid}
\title{I.I.D. Random Sketching Estimator}
\usage{
ske_iid(X, y, m, c, type = 1, partial = 0)
}
\arguments{
\item{X}{Design matrix (n x p)}

\item{y}{Response vector (n x 1)}

\item{m}{Sketch size (number of rows in sketched matrix)}

\item{c}{Linear combination vector (p x 1)}

\item{type}{Distribution type: 1=Gaussian, 2=Three-point, 3=Sparse three-point (default=1)}

\item{partial}{Whether to use partial sketching (0=full sketching, 1=partial sketching)}
}
\value{
List containing:
\item{sketched_ls_sol}{Sketched least squares solution}
\item{linear_combination}{c'*sketched_ls_sol}
\item{sketched_data}{Sketched design matrix and response}
}
\description{
Performs I.I.D. random sketching for least squares problem using various
random distributions.
}
\examples{
# Generate sample data
set.seed(123)
n <- 100; p <- 5; m <- 30
X <- matrix(rnorm(n * p), n, p)
y <- X \%*\% c(1, -0.5, 0.3, 0.8, -0.2) + rnorm(n, sd = 0.1)
c_vec <- rep(1, p)  # sum of coefficients

# Perform I.I.D. sketching
result <- ske_iid(X, y, m, c_vec)
print(result$sketched_ls_sol)
print(result$linear_combination)

}
